@@ -1,12 +1,15 @@
 package ht.treechop.client.gui.element;
 
-import com.google.common.collect.Lists;
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
+import ht.treechop.client.gui.util.GUIUtil;
 import ht.treechop.client.gui.widget.StickyWidget;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.client.gui.widget.Widget;
-import net.minecraft.util.text.ITextComponent;
+import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
@@ -15,21 +18,24 @@ import java.util.stream.Collectors;
 
 public class ExclusiveButtonsGui extends NestedGui {
 
-    private final List<Widget> widgets;
+    private final List<AbstractWidget> widgets;
+    private final Supplier<Component> tooltipSupplier;
 
-    protected ExclusiveButtonsGui(Collection<Widget> widgets) {
-        this.widgets = Lists.newArrayList(widgets);
+    protected ExclusiveButtonsGui(Collection<AbstractWidget> widgets, Supplier<Component> tooltipSupplier) {
+        super(0, 0, 0, 0, TextComponent.EMPTY);
+        this.widgets = new ArrayList<>(widgets);
+        this.tooltipSupplier = tooltipSupplier;
     }
 
-    public void resize(int width) {
+    public void expand(int width) {
         if (getMinimumWidth() < width) {
             int targetWidth = width / widgets.size();
-            List<Widget> smallerWidgets = widgets.stream().filter(widget -> widget.getWidth() <= targetWidth).collect(Collectors.toList());
-            List<Widget> biggerWidgets = widgets.stream().filter(widget -> widget.getWidth() > targetWidth).collect(Collectors.toList());
-            int totalWidthForSmallers = width - biggerWidgets.stream().map(Widget::getWidth).reduce(Integer::sum).orElse(0);
+            List<AbstractWidget> smallerWidgets = widgets.stream().filter(widget -> widget.getWidth() <= targetWidth).collect(Collectors.toList());
+            List<AbstractWidget> biggerWidgets = widgets.stream().filter(widget -> widget.getWidth() > targetWidth).collect(Collectors.toList());
+            int totalWidthForSmallers = width - biggerWidgets.stream().map(AbstractWidget::getWidth).reduce(Integer::sum).orElse(0);
             int i = 0;
             // Do this incrementally to account for rounding errors
-            for (Widget widget : smallerWidgets) {
+            for (AbstractWidget widget : smallerWidgets) {
                 double lower = (double) i / smallerWidgets.size();
                 double upper = (double) (i + 1) / smallerWidgets.size();
                 int widgetWidth = (int) (totalWidthForSmallers * (upper - lower));
@@ -41,53 +47,69 @@ public class ExclusiveButtonsGui extends NestedGui {
 
     @SuppressWarnings("NullableProblems")
     @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
+    public List<? extends GuiEventListener> children() {
         return widgets;
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         int x = getBox().getLeft();
         int y = getBox().getTop();
-        for (Widget widget : widgets) {
-            widget.x = x;
+
+        int maxX = x;
+        int maxY = y;
+
+        for (AbstractWidget widget : widgets) {
+            widget.x = maxX;
             widget.y = y;
-            widget.render(matrixStack, mouseX, mouseY, partialTicks);
-            x += widget.getWidth();
+            widget.render(poseStack, mouseX, mouseY, partialTicks);
+
+            maxX = Math.max(maxX, maxX + widget.getWidth());
+            maxY = Math.max(maxY, y + widget.getHeight());
+        }
+
+        isHovered = mouseX >= x && mouseY >= y && mouseX < maxX && mouseY < maxY;
+        if (isHoveredOrFocused()) {
+            GUIUtil.showTooltip(mouseX, mouseY, tooltipSupplier.get());
         }
     }
 
     @Override
     public int getMinimumWidth() {
-        return widgets.stream().map(Widget::getWidth).reduce(Integer::sum).orElse(0);
+        return widgets.stream().map(AbstractWidget::getWidth).reduce(Integer::sum).orElse(0);
     }
 
     @Override
     public int getMinimumHeight() {
-        return widgets.stream().map(Widget::getHeightRealms).reduce(Integer::max).orElse(0);
+        return widgets.stream().map(AbstractWidget::getHeight).reduce(Integer::max).orElse(0);
+    }
+
+    @Override
+    public void updateNarration(NarrationElementOutput out) {
+        // TODO
     }
 
     public static class Builder {
         private final List<Option> options = new LinkedList<>();
 
-        public Builder add(ITextComponent name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
+        public Builder add(Component name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
             options.add(new Option(name, onPress, stateSupplier));
             return this;
         }
 
-        public ExclusiveButtonsGui build() {
-            List<Widget> widgets = options.stream()
+        public ExclusiveButtonsGui build(Supplier<Component> tooltipSupplier) {
+            List<AbstractWidget> widgets = options.stream()
                     .map(option -> new StickyWidget(0, 0, 0, 0, option.name, option.onPress, option.stateSupplier))
                     .collect(Collectors.toList());
-            return new ExclusiveButtonsGui(widgets);
+            return new ExclusiveButtonsGui(widgets, tooltipSupplier);
         }
 
         private static class Option {
-            private ITextComponent name;
-            private Runnable onPress;
-            private Supplier<StickyWidget.State> stateSupplier;
+            private final Component name;
+            private final Runnable onPress;
+            private final Supplier<StickyWidget.State> stateSupplier;
 
-            public Option(ITextComponent name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
+            public Option(Component name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
                 this.name = name;
                 this.onPress = onPress;
                 this.stateSupplier = stateSupplier;
