@@ -1,10 +1,11 @@
 package ht.treechop.client.gui.element;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
 import ht.treechop.client.gui.widget.TextWidget;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.util.text.ITextComponent;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.Component;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -22,7 +23,8 @@ public class LabeledGui extends NestedGui {
     private int rightcolumnWidth;
     private boolean rightAlignLabels = false;
 
-    public LabeledGui(FontRenderer font, ITextComponent label, NestedGui gui) {
+    public LabeledGui(Font font, Component label, NestedGui gui) {
+        super(0, 0, 0, 0, label);
         this.label = new TextWidget(0, 0, font, label);
         this.gui = gui;
         this.leftColumnWidth = getLeftColumnWidth();
@@ -34,7 +36,7 @@ public class LabeledGui extends NestedGui {
     }
 
     public int getMinimumHeight() {
-        return Math.max(label.getHeightRealms(), gui.getMinimumHeight());
+        return Math.max(label.getHeight(), gui.getMinimumHeight());
     }
 
     public int getLeftColumnWidth() {
@@ -46,21 +48,21 @@ public class LabeledGui extends NestedGui {
     }
 
     @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
-        return Stream.of(Collections.singletonList(label), gui.getEventListeners())
+    public List<? extends GuiEventListener> children() {
+        return Stream.of(Collections.singletonList(label), gui.children())
                 .flatMap(Collection::stream)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         int center = getBox().getCenterX() + (leftColumnWidth - rightcolumnWidth) / 2;
         this.label.x = center - COLUMN_PADDING + (rightAlignLabels ? 0 : -leftColumnWidth);
         this.label.y = getBox().getCenterY() - 3;
-        this.label.render(matrixStack, mouseX, mouseY, partialTicks, rightAlignLabels);
+        this.label.render(poseStack, mouseX, mouseY, partialTicks, rightAlignLabels);
         this.gui.setBox(center + COLUMN_PADDING, getBox().getTop(), rightcolumnWidth, getBox().getHeight());
         this.gui.render(
-                matrixStack,
+                poseStack,
                 mouseX,
                 mouseY,
                 partialTicks
@@ -70,6 +72,11 @@ public class LabeledGui extends NestedGui {
     public void setColumnWidths(int biggestLeftColumnWidth, int biggestRightColumnWidth) {
         this.leftColumnWidth = biggestLeftColumnWidth;
         this.rightcolumnWidth = biggestRightColumnWidth;
-//        element.resize(biggestRightColumnWidth);
+        gui.expand(biggestRightColumnWidth);
+    }
+
+    @Override
+    public void updateNarration(NarrationElementOutput p_169152_) {
+        // TODO
     }
 }
