@@ -1,8 +1,9 @@
 package ht.treechop.client.gui.element;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-import ht.treechop.client.gui.util.ScreenBox;
-import net.minecraft.client.gui.IGuiEventListener;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.TextComponent;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -14,9 +15,9 @@ public class RowsGui extends NestedGui {
     private final int biggestLeftColumnWidth;
     private final int biggestRightColumnWidth;
     private final List<NestedGui> rows = new ArrayList<>();
-    private ScreenBox box;
 
     public RowsGui(int rowSeparation, Collection<NestedGui> rows) {
+        super(0, 0, 0, 0, TextComponent.EMPTY);
         this.rowSeparation = rowSeparation;
         this.rows.addAll(rows);
 
@@ -35,11 +36,11 @@ public class RowsGui extends NestedGui {
 
     @SuppressWarnings("NullableProblems")
     @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
+    public List<? extends GuiEventListener> children() {
         return rows;
     }
 
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         int rowLeft = getBox().getLeft();
         int rowTop = getBox().getTop();
         int rowWidth = getBox().getWidth();
@@ -47,7 +48,7 @@ public class RowsGui extends NestedGui {
         for (NestedGui row : rows) {
             int rowHeight = row.getMinimumHeight();
             row.setBox(rowLeft, rowTop, rowWidth, rowHeight);
-            row.render(matrixStack, mouseX, mouseY, partialTicks);
+            row.render(poseStack, mouseX, mouseY, partialTicks);
             rowTop += rowHeight + rowSeparation;
         }
     }
@@ -63,12 +64,7 @@ public class RowsGui extends NestedGui {
     }
 
     @Override
-    public ScreenBox getBox() {
-        return this.box;
-    }
-
-    @Override
-    public void setBox(ScreenBox box) {
-        this.box = box;
+    public void updateNarration(NarrationElementOutput p_169152_) {
+        // TODO
     }
 }
