@@ -1,9 +1,13 @@
 package ht.treechop.client.gui.element;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
+import ht.treechop.client.gui.util.GUIUtil;
 import ht.treechop.client.gui.widget.ToggleWidget;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.client.gui.widget.Widget;
+import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
 
 import java.util.Collections;
 import java.util.List;
@@ -11,23 +15,29 @@ import java.util.function.Supplier;
 
 public class ToggleGui extends NestedGui {
 
-    private final Widget widget;
+    private final AbstractWidget widget;
+    private final Supplier<Component> tooltipSupplier;
 
-    public ToggleGui(Runnable onPress, Supplier<ToggleWidget.State> stateSupplier) {
+    public ToggleGui(Runnable onPress, Supplier<ToggleWidget.State> stateSupplier, Supplier<Component> componentSupplier) {
+        super(0, 0, 0, 0, TextComponent.EMPTY);
         this.widget = new ToggleWidget(0, 0, onPress, stateSupplier);
+        this.tooltipSupplier = componentSupplier;
     }
 
-    @SuppressWarnings("NullableProblems")
     @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
+    public List<? extends GuiEventListener> children() {
         return Collections.singletonList(widget);
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         widget.x = getBox().getLeft();
         widget.y = getBox().getTop();
-        widget.render(matrixStack, mouseX, mouseY, partialTicks);
+        widget.render(poseStack, mouseX, mouseY, partialTicks);
+
+        if (widget.isHoveredOrFocused()) {
+            GUIUtil.showTooltip(mouseX, mouseY, tooltipSupplier.get());
+        }
     }
 
     @Override
@@ -37,7 +47,11 @@ public class ToggleGui extends NestedGui {
 
     @Override
     public int getMinimumHeight() {
-        return widget.getHeightRealms();
+        return widget.getHeight();
     }
 
+    @Override
+    public void updateNarration(NarrationElementOutput p_169152_) {
+        // TODO
+    }
 }
