@@ -1,26 +1,19 @@
 package ht.treechop.client.gui.widget;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import ht.treechop.client.gui.util.Sprite;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.widget.Widget;
-import net.minecraft.util.text.StringTextComponent;
-import org.apache.commons.lang3.tuple.Pair;
+import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.TextComponent;
 
-import java.util.EnumMap;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-public class SpriteButtonWidget extends Widget {
+public class SpriteButtonWidget extends AbstractWidget {
 
     private final Runnable onPress;
     private final Sprite sprite;
     private final Sprite highlightedSprite;
 
     public SpriteButtonWidget(int x, int y, Sprite sprite, Sprite highlightedSprite, Runnable onPress) {
-        super(x, y, Math.max(sprite.width, highlightedSprite.width), Math.max(sprite.height, highlightedSprite.height), new StringTextComponent(""));
+        super(x, y, Math.max(sprite.width, highlightedSprite.width), Math.max(sprite.height, highlightedSprite.height), new TextComponent(""));
         this.onPress = onPress;
         this.sprite = sprite;
         this.highlightedSprite = highlightedSprite;
@@ -29,8 +22,8 @@ public class SpriteButtonWidget extends Widget {
 
     @SuppressWarnings("NullableProblems")
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
-        super.render(matrixStack, mouseX, mouseY, partialTicks);
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
+        super.render(poseStack, mouseX, mouseY, partialTicks);
     }
 
     public void onClick(double mouseX, double mouseY) {
@@ -39,10 +32,14 @@ public class SpriteButtonWidget extends Widget {
 
     @SuppressWarnings("NullableProblems")
     @Override
-    public void renderButton(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void renderButton(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         Sprite.setRenderState(this.alpha);
-        Sprite sprite = isHovered() ? highlightedSprite : this.sprite;
-        sprite.blit(matrixStack, x, y);
+        Sprite sprite = isHoveredOrFocused() ? highlightedSprite : this.sprite;
+        sprite.blit(poseStack, x, y);
     }
 
+    @Override
+    public void updateNarration(NarrationElementOutput out) {
+        // TODO
+    }
 }
