@@ -1,33 +1,38 @@
 package ht.treechop.client.gui.widget;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
 import ht.treechop.client.gui.util.GUIUtil;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.gui.widget.Widget;
-import net.minecraft.util.text.ITextComponent;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.Component;
 
-public class TextWidget extends Widget {
+public class TextWidget extends AbstractWidget {
 
-    private FontRenderer font;
+    private Font font;
 
-    public TextWidget(int x, int y, FontRenderer font, ITextComponent text) {
-        super(x, y, font.getStringWidth(text.getString()), GUIUtil.TEXT_LINE_HEIGHT, text);
+    public TextWidget(int x, int y, Font font, Component text) {
+        super(x, y, font.width(text.getString()), GUIUtil.TEXT_LINE_HEIGHT, text);
         this.font = font;
     }
 
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks, boolean rightAligned) {
-        render(matrixStack, mouseX, mouseY, partialTicks, rightAligned ? -font.getStringWidth(getMessage().getString()) : 0);
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks, boolean rightAligned) {
+        render(poseStack, mouseX, mouseY, partialTicks, rightAligned ? -font.width(getMessage()) : 0);
     }
 
     @SuppressWarnings("NullableProblems")
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
-        render(matrixStack, mouseX, mouseY, partialTicks, 0);
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
+        render(poseStack, mouseX, mouseY, partialTicks, 0);
     }
 
     @SuppressWarnings({"SuspiciousNameCombination"})
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks, int xOffset) {
-        drawString(matrixStack, font, getMessage(), x + xOffset, y, 0xFFFFFF);
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks, int xOffset) {
+        drawString(poseStack, font, getMessage(), x + xOffset, y, 0xFFFFFF);
     }
 
+    @Override
+    public void updateNarration(NarrationElementOutput out) {
+        // TODO
+    }
 }
