@@ -2,9 +2,10 @@ package ht.treechop.common.compat;
 
 import harmonised.pmmo.events.BlockBrokenHandler;
 import ht.treechop.TreeChopMod;
+import ht.treechop.api.ChopEvent;
+import ht.treechop.common.block.ChoppedLogBlock;
 import ht.treechop.common.config.ConfigHandler;
-import ht.treechop.common.event.ChopEvent;
-import net.minecraft.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -25,10 +26,16 @@ public class ProjectMMO {
     private static class EventHandler {
         @SubscribeEvent
         public static void onFinishChop(ChopEvent.FinishChopEvent event) {
+            BlockState xpBlock;
+            if (event.getWorld().getBlockEntity(event.getChoppedBlockPos()) instanceof ChoppedLogBlock.Entity entity) {
+                xpBlock = entity.getOriginalState();
+            } else {
+                xpBlock = event.getChoppedBlockState();
+            }
             BlockBrokenHandler.handleBroken(new BlockEvent.BreakEvent(
                     event.getWorld(),
                     event.getChoppedBlockPos(),
-                    Blocks.OAK_LOG.getDefaultState(), // TODO: use correct wood type
+                    xpBlock,
                     event.getPlayer()
             ));
         }
