@@ -1,11 +1,11 @@
 package ht.treechop.common.config.item;
 
-import net.minecraft.item.Item;
-import net.minecraft.tags.ITagCollection;
+import net.minecraft.tags.TagCollection;
+import net.minecraft.world.item.Item;
 import net.minecraftforge.registries.IForgeRegistry;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class ItemNamespaceIdentifier extends ItemIdentifier {
 
@@ -14,14 +14,9 @@ public class ItemNamespaceIdentifier extends ItemIdentifier {
     }
 
     @Override
-    public List<Item> resolve(ITagCollection<Item> tags, IForgeRegistry<Item> registry) {
-        List<Item> items = registry.getValues().stream()
-                .filter(item -> item.getRegistryName() != null && item.getRegistryName().getNamespace().equals(getNamespace()))
-                .collect(Collectors.toList());
-        if (items.isEmpty()) {
-            parsingError(String.format("no items found in namespace \"%s\"", getNamespace()));
-        }
-        return items;
+    public Stream<Item> resolve(TagCollection<Item> tags, IForgeRegistry<Item> registry) {
+        return registry.getValues().stream()
+                .filter(item -> item.getRegistryName() != null && item.getRegistryName().getNamespace().equals(getNamespace()));
     }
 
 }
