@@ -1,11 +1,11 @@
 package ht.treechop.common.config.item;
 
-import net.minecraft.item.Item;
-import net.minecraft.tags.ITagCollection;
+import net.minecraft.tags.TagCollection;
+import net.minecraft.world.item.Item;
 import net.minecraftforge.registries.IForgeRegistry;
 
 import java.util.Collections;
-import java.util.List;
+import java.util.stream.Stream;
 
 public class MalformedItemIdentifier extends ItemIdentifier {
 
@@ -15,8 +15,8 @@ public class MalformedItemIdentifier extends ItemIdentifier {
     }
 
     @Override
-    public List<Item> resolve(ITagCollection<Item> tags, IForgeRegistry<Item> registry) {
-        return Collections.emptyList();
+    public Stream<Item> resolve(TagCollection<Item> tags, IForgeRegistry<Item> registry) {
+        return Stream.empty();
     }
 
 }
